decouple rehydration component from water component
