+++ src/main/resources/fabric.mod.json
@@ -51,7 +51,8 @@
 	"custom": {
 		"cardinal-components": [
 			"scorchful:player_water",
-			"scorchful:entity_desert_vision"
+			"scorchful:entity_desert_vision",
+			"scorchful:rehydration"
 		],
 		"modmenu": {
 			"links": {
